Pull request #3536: US #762655 detection: target service http rules to specific message sections - Part 5
Merge in SNORT/snort3 from ~MDAGON/snort3:proto_5 to master

Squashed commit of the following:

commit 83ef46f4c04816c433d40af59cda244aaacde1b2
Author: Tom Peters <thopeter@cisco.com>
Date:   Mon Mar 21 16:39:24 2022 -0400

    http_inspect: remove rule option timing features

+++ src/framework/mpse_batch.cc
@@ -72,12 +72,14 @@ MpseGroup::~MpseGroup()
 {
     if (normal_mpse)
     {
-        MpseManager::delete_search_engine(normal_mpse);
+        if (!normal_is_dup)
+            MpseManager::delete_search_engine(normal_mpse);
         normal_mpse = nullptr;
     }
     if (offload_mpse)
     {
-        MpseManager::delete_search_engine(offload_mpse);
+        if (!offload_is_dup)
+            MpseManager::delete_search_engine(offload_mpse);
         offload_mpse = nullptr;
     }
 }
